requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
transformers==4.35.2
torch==2.1.2
huggingface-hub==0.19.4
accelerate==0.24.1
streamlit==1.31.0
anthropic>=0.9.0 